examples/quickstart-pytorch-slurm/requirements.txt
flwr>=1.0, <2.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
hydra-core==1.3.2
